+++ lib/gssapi/mech/gss_import_name.c
@@ -132,7 +132,7 @@ _gss_import_export_name(OM_uint32 *minor_status,
                 p += t;
                 len -= t;
 
-                t = (p[0] << 24) | (p[1] << 16) | (p[2] << 8) | p[3];
+                t = ((unsigned long)p[0] << 24) | (p[1] << 16) | (p[2] << 8) | p[3];
                 /* p += 4; // we're done using `p' now */
                 len -= 4;
